Added Ardèche as test site. All OK except observations_search (TBD)

+++ export_vn/test/test_download_vn.py
@@ -23,8 +23,8 @@ def test_logging(cmdopt, capsys):
             logging.getLogger().setLevel(logging.DEBUG)
         logging.debug('Running with debug logging level')
 
-# Using t38 site, that needs to be created first
-SITE = 't38'
+# Using t07 site, that needs to be created first
+SITE = 't07'
 
 def test_site():
     """Check if configuration file exists."""
@@ -44,20 +44,19 @@ TERRITORIAL_UNIT = TerritorialUnits(CFG)
 # -----------------
 #  LocalAdminUnits
 # -----------------
-def test_local_admin_units_init(capsys):
-    """Get from template: a single local admin unit."""
-    assert LOCAL_ADMIN_UNITS.transfer_errors == 0
-
 def test_local_admin_units_store(capsys):
     """Store local_admin_units to file."""
     file_json = str(Path.home()) + '/' + CFG.file_store + 'local_admin_units_1.json.gz'
-    Path(file_json).unlink()
+    if Path(file_json).is_file():
+        Path(file_json).unlink()
     LOCAL_ADMIN_UNITS.store()
     assert Path(file_json).is_file()
     with gzip.open(file_json, 'rb') as g:
         items_dict = json.loads(g.read().decode('utf-8'))
-    assert len(items_dict['data']) > 530
-
+    if SITE == 't38':
+        assert len(items_dict['data']) >= 534
+    elif SITE == 't07':
+        assert len(items_dict['data']) >= 340
 
 # -------------
 #  Observations
@@ -65,7 +64,8 @@ def test_local_admin_units_store(capsys):
 def test_observations_store_l_18(capsys):
     """Store observations to file, using list."""
     file_json = str(Path.home()) + '/' + CFG.file_store + 'observations_18_1.json.gz'
-    Path(file_json).unlink()
+    if Path(file_json).is_file():
+        Path(file_json).unlink()
     OBSERVATIONS.store(18, method='list')
     assert Path(file_json).is_file()
     with gzip.open(file_json, 'rb') as g:
@@ -79,12 +79,16 @@ def test_observations_store_l_18(capsys):
 def test_places_store(capsys):
     """Store places to file."""
     file_json = str(Path.home()) + '/' + CFG.file_store + 'places_1.json.gz'
-    Path(file_json).unlink()
+    if Path(file_json).is_file():
+        Path(file_json).unlink()
     PLACES.store()
     assert Path(file_json).is_file()
     with gzip.open(file_json, 'rb') as g:
         items_dict = json.loads(g.read().decode('utf-8'))
-    assert len(items_dict['data']) > 31930
+    if SITE == 't38':
+        assert len(items_dict['data']) >= 31930
+    elif SITE == 't07':
+        assert len(items_dict['data']) >= 23566
 
 # --------
 #  Species
@@ -92,7 +96,8 @@ def test_places_store(capsys):
 def test_species_store(capsys):
     """Store species to file."""
     file_json = str(Path.home()) + '/' + CFG.file_store + 'species_1.json.gz'
-    Path(file_json).unlink()
+    if Path(file_json).is_file():
+        Path(file_json).unlink()
     SPECIES.store()
     assert Path(file_json).is_file()
     with gzip.open(file_json, 'rb') as g:
@@ -105,7 +110,8 @@ def test_species_store(capsys):
 def test_taxo_groups_store(capsys):
     """Store taxo groups to file."""
     file_json = str(Path.home()) + '/' + CFG.file_store + 'taxo_groups_1.json.gz'
-    Path(file_json).unlink()
+    if Path(file_json).is_file():
+        Path(file_json).unlink()
     TAXO_GROUP.store()
     assert Path(file_json).is_file()
     with gzip.open(file_json, 'rb') as g:
@@ -118,7 +124,8 @@ def test_taxo_groups_store(capsys):
 def test_territorial_units_store(capsys):
     """Store territorial units to file."""
     file_json = str(Path.home()) + '/' + CFG.file_store + 'territorial_units_1.json.gz'
-    Path(file_json).unlink()
+    if Path(file_json).is_file():
+        Path(file_json).unlink()
     TERRITORIAL_UNIT.store()
     assert Path(file_json).is_file()
     with gzip.open(file_json, 'rb') as g:
